refactor(settings): segregate app settings into Appearance / Dashboard / General
The single "Appearance" panel had grown into a grab-bag mixing three kinds of
setting — cosmetic (theme/color/interface), behavioral (live chart window), and
informational (storage paths). Split it into three focused client categories and
model the app side as a data-driven registry, mirroring how the engine settings
side already works.

- app-panels registry: declares client panels (id, label, icon, description,
  Component) as data. SettingsMain looks up the panel instead of branching on a
  category string; the sidebar tree renders labels/icons from the same source.
  Adding an app category is now one entry + one panel file.
- Split UISettingsPanel into AppearancePanel (theme/color/interface),
  DashboardPanel (live chart window — its honest home), and GeneralPanel
  (storage paths, with a non-Electron fallback), sharing a ClientSettingsPanel
  shell for the header + scroll chrome.
- Category rows drop the six bespoke per-category colors for a single
  centralized `--primary` accent on the selected row, consistent with the rest
  of the app (RunItem/RequestItem).
- Types: split SettingsCategory into ClientSettingsCategory + EngineSettingsCategory;
  SettingsMain's save-context guard now keys off isClientCategory rather than the
  literal "ui" so client panels never get the engine Save/Reset bar.
- Remove the uPlot chart spike docs/assets now that the work has shipped.
